requirements.txt
Flask==3.0.3
Jinja2==3.1.6
python-dotenv==1.0.1
Authlib==1.3.2
supabase==2.6.0
httpx==0.27.2
itsdangerous==2.2.0
werkzeug==3.1.3
APScheduler==3.10.1
